# src/albums/dto/create-album.dto.ts
import { IsDefined, IsInt, IsOptional, IsUUID } from 'class-validator';
class CreateAlbumDto {
  @IsDefined()
  name: string;
  @IsDefined()
  @IsInt()
  year: number;
  @IsOptional()
  @IsUUID()
  artistId?: string | null;
}

export default CreateAlbumDto;